requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework==3.11.0
drf-extensions==0.6.0
requests==2.24.0